% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotwidget_update.R
\name{get_afterchange}
\alias{get_afterchange}
\title{Get and parse the changes logged by the afterChange hook}
\usage{
get_afterchange(input, hotwidget_data)
}
\arguments{
\item{input}{The input list from the server function}

\item{hotwidget_data}{The original data set}
}
\value{
A list with the data set and a boolean indicating if an error occured.
}
\description{
Get and parse the changes logged by the afterChange hook
}
