% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotwidget_update.R
\name{hotwidget_update}
\alias{hotwidget_update}
\title{hotwidget_update}
\usage{
hotwidget_update(input, id, hotwidget_data_updated, verbose = FALSE)
}
\arguments{
\item{input}{the list of input from the server function}

\item{id}{the id of the hotwidget}

\item{hotwidget_data_updated}{a reactiveVal(), initiated with the original
data set. This object will contain the data which has been updated by the
user.}

\item{verbose}{logical, if TRUE print debug messages}
}
\value{
Doesnt return anything, but as a side effect updates
hotwidget_data_updated.
}
\description{
Updates the data set with the changes made by the user in the hotwidget.
It works through collecting the changes made by the user via different
hooks in the handsontable.js library.
\itemize{
\item afterChange
\item afterCreateRow
\item afterRemoveRow
\item afterCreateCol
\item afterRemoveCol
\item afterUndo
\item afterRedo
}

Data types are inferred via the original data set. This leads to the
following behaviour
\itemize{
\item non-numerics in numeric columns become NA
\item invalid dates are become NA
\item factors where the value is not in the levels become NA
}

known issues:
\href{https://github.com/handsontable/handsontable/issues/10610}{GitHub issue}
\itemize{
\item combination of sorting and undo/redo doesnt function properly.
\item hotwidget disables undo/redo if you want column sorting
}
}
