% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotwidget.R
\name{hotwidget}
\alias{hotwidget}
\title{hotwidget: Handsontable widget for R}
\usage{
hotwidget(
  data,
  columns = NULL,
  width = NULL,
  height = NULL,
  rowHeaders = FALSE,
  columnSorting = TRUE,
  allowRemoveRow = TRUE,
  allowInsertRow = TRUE,
  autoWrapRow = TRUE,
  autoWrapCol = TRUE,
  filters = TRUE,
  dropdownMenu = TRUE,
  contextMenu = TRUE,
  licenseKey = NA
)
}
\arguments{
\item{data}{A data frame or matrix}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{rowHeaders}{Show row headers}

\item{columnSorting}{Enable column sorting}

\item{autoWrapRow}{Enable auto wrap for rows}

\item{autoWrapCol}{Enable auto wrap for columns}

\item{filters}{Enable filters}

\item{dropdownMenu}{Enable dropdown menu}

\item{contextMenu}{Enable context menu}

\item{licenseKey}{License key}

\item{colHeaders}{Show column headers}
}
\description{
hotwidget: Handsontable widget for R
}
