% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotwidget.R
\name{hotwidget}
\alias{hotwidget}
\title{hotwidget: Handsontable widget for R}
\usage{
hotwidget(
  data,
  key_column = NULL,
  columns = NULL,
  hiddenColumns = list(),
  width = NULL,
  height = NULL,
  rowHeaders = FALSE,
  undo = TRUE,
  columnSorting = TRUE,
  allowRemoveRow = TRUE,
  allowInsertRow = TRUE,
  autoWrapRow = TRUE,
  autoWrapCol = TRUE,
  filters = TRUE,
  dropdownMenu = TRUE,
  contextMenu = TRUE,
  licenseKey = NA,
  ...
)
}
\arguments{
\item{data}{A data frame or matrix}

\item{key_column}{Key column to identify rows (has to be integer), client will be updated with max + 1 for new rows}

\item{columns}{Column definition, if not assigned will be guessed
e.g.:
\preformatted{
list(
  list(
    data = "{column_name}",
    type = "data type",
    source = c("choices", "for", "dropdown"),
    numericFormat = list(
      pattern = "0.00"
    ),
    dateFormat = "YYYY-MM-DD"
  ),
  list(...)
)
}
More details at: \href{https://handsontable.com/docs/javascript-data-grid/cell-type/}{Documentation Handsontable.js}}

\item{hiddenColumns}{Default is enabled, to specify use list(columns = list(0,1,2), indicators = TRUE, copyPasteEnabled = FALSE)}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{rowHeaders}{Show row headers}

\item{undo}{Enable undo/redo functionality, if columnSorting is enabled, undo/redo is disabled}

\item{columnSorting}{Enable column sorting (if you enable this, undo/redo functionality is disabled)}

\item{allowRemoveRow}{Allow the removal of rows}

\item{allowInsertRow}{Allow row inserting}

\item{autoWrapRow}{Enable auto wrap for rows}

\item{autoWrapCol}{Enable auto wrap for columns}

\item{filters}{Enable filters}

\item{dropdownMenu}{Enable dropdown menu}

\item{contextMenu}{Enable context menu}

\item{licenseKey}{License key}
}
\description{
known issues:
\itemize{
\item combination of sorting and undo/redo doesnt function properly.
\item hotwidget disables undo/redo if you want column sorting
}
}
