% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Test app for debugging, test and possibly example purposes}
\usage{
run_app(
  hotwidget_data = head(mutate(janitor::clean_names(iris), .before = 1, index =
    row_number(), test = as_date(paste(Sys.Date()))), 6)
)
}
\arguments{
\item{hotwidget_data}{data set}
}
\value{
app
}
\description{
Test app for debugging, test and possibly example purposes
}
